requirements.txt
Flask
flask-cors
flask-pymongo
scapy
speedtest-cli
apscheduler
psutil
pymongo
